% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecat_and_scaling_functions.R
\name{calculate_ecat}
\alias{calculate_ecat}
\title{Calculate ECAT exposure estimates at specific locations.}
\usage{
calculate_ecat(locations, return.LU.vars = FALSE)
}
\arguments{
\item{locations}{Data.frame with columns 'id', 'lat', and 'lon' at minimum.}

\item{return.LU.vars}{When \code{return.LU.vars = TRUE}, the land use predictors used
to generate the ECAT values are also returned.}
}
\value{
If \code{return.LU.vars = FALSE}, a numeric vector of ECAT estimates (ug/m3)
    is returned. If \code{return.LU.vars = TRUE}, the \code{locations} data.frame with
    additional columns for ECAT values and the land use predictors used
    to generate the ECAT values is returned.
}
\description{
\code{calculate_ecat()} uses a land use regression model developed by Dr. Patrick Ryan
    based on ambient air sampling in Cincinnati, OH between 2001 and 2005 to estimate
    exposure to elemental carbon attributable to traffic (ECAT) at point locations in
    the area specified by latitude and longitude. The model predictors include elevation,
    truck traffic within 400 meters, and length of bus routes within 100 meters. Returned
    ECAT values are in micrograms per cubic meter.
}
\examples{
my_data <- data.frame(id = 1:3,
    lat = c(39.19674, 39.12731,	39.28765),
    lon = c(-84.58260, -84.52700, -84.51017))

ecat_est <- calculate_ecat(my_data, return.LU.vars = FALSE)
ecat_est <- calculate_ecat(my_data, return.LU.vars = TRUE)
}
\references{
Ryan, P.H., G.K. LeMasters, P. Biswas, L. Levin, S. Hu, M. Lindsey, D.I. Bernstein, J. Lockey, M. Villareal,
G.K. Khurana Hershey, and S.A. Grinshpun. 2007. "A Comparison of Proximity and Land Use Regression Traffic Exposure Models
and Wheezing in Infants." Environmental Health Perspectives 115(2): 278-284.  \url{https://doi.org/10.1289/ehp.9480}
}
