% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecat_and_scaling_functions.R
\name{calculate_scaled_ecat}
\alias{calculate_scaled_ecat}
\title{Calculate temporally scaled ECAT exposure estimates at specific locations.}
\usage{
calculate_scaled_ecat(locations, days_prior)
}
\arguments{
\item{locations}{Data.frame with columns 'id', 'lat','lon', and 'date' at minimum.}

\item{days_prior}{The number of days prior to each date to average over. Must
be an integer.}
}
\value{
a numeric vector of ECAT estimates (ug/m3).
}
\description{
\code{calculate_scaled_ecat()} is a wrapper function that estimates ECAT exposures
    at provided locations by calling \code{calculate_ecat()}, then temporally scales those
    estimates using scaling factors computed by calling \code{calculate_scaling_factors()}.
    This function is particularly useful for calculating exposures at the same locations on
    different dates.
}
\examples{
my_data <- data.frame(id = rep(1,3),
    lat = c(39.19674, 39.19674,	39.19674),
    lon = c(-84.58260, -84.58260, -84.58260),
    date = c("2010-01-08", "2012-06-08", "2015-04-09"))

ecat_scaled <- calculate_scaled_ecat(my_data, months_prior = 1)
}
