% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecat_and_scaling_functions.R
\name{calculate_scaling_factors}
\alias{calculate_scaling_factors}
\title{Calculate temporal scaling factors based on EPA measurements of EC.}
\usage{
calculate_scaling_factors(dates, days_prior)
}
\arguments{
\item{dates}{Vector of class \code{Date}. See \code{\link{as.Date}} for help converting
a character vector to a Date vector.}

\item{days_prior}{The number of days prior to each date to average over. Must
be an integer.}
}
\value{
A numeric vector of temporal scaling factors the same length
    as \code{dates}.
}
\description{
\code{calculate_scaling_factors()} constructs temporal scaling factors based on measurements
    of elemental carbon (EC) recorded by the EPA in the Cincinnati area. These scaling factors are
    the average EC measured over the provided number of \code{days_prior} up to the provided
    \code{dates}, divided by the average EC recorded over the ECAT ambient air sampling period (2001 to 2005).
    Scaling factors can be multiplied by ECAT estimates from \code{calculate_ecat()} to adjust for
    temporal variability in ECAT in the Cincinnati area over time.
}
\details{
EPA data in this package is available from November 9, 2001
    through November 28, 2018. Scaling factors that attempt to average over
    EC measured on dates outside this range will not be calculated. In addition,
    it is important to be mindful of the frequency of EC measurements recorded by the EPA
    when choosing a value for \code{days_prior}. Note that EC was measured every 6 days
    through the end of 2010, and every 3 days starting in 2011.
}
\examples{
my_dates <- c("2010-01-08", "2012-06-08", "2010-01-09", "2015-04-09", "2010-01-10")

\dontrun{
class(my_dates)  # character vector
scaling1m <- calculate_scaling_factors(my_dates, days_prior = 30)
}

my_dates <- as.Date(my_dates)
class(my_dates)  # Date vector
scaling1m <- calculate_scaling_factors(my_dates, days_prior = 30)
}
