% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecat_and_scaling_functions.R
\name{calculate_scaled_ecat}
\alias{calculate_scaled_ecat}
\title{Calculate temporally scaled ECAT exposure estimates at specific locations.}
\usage{
calculate_scaled_ecat(locations)
}
\arguments{
\item{locations}{Data.frame with columns 'id', 'lat','lon', 'start_date', and 'end_date' at minimum.}
}
\value{
A numeric vector of ECAT estimates (ug/m3).
}
\description{
\code{calculate_scaled_ecat()} is a wrapper function that estimates ECAT exposures
    at provided locations by calling \code{calculate_ecat()}, then temporally scales those
    estimates using scaling factors computed by calling \code{calculate_scaling_factors()}.
    This function is particularly useful for calculating exposures at the same locations on
    different dates.
}
\examples{
my_data <- data.frame(id = rep(1,3),
    lat = c(39.19674, 39.19674,	39.19674),
    lon = c(-84.58260, -84.58260, -84.58260),
    start_date = c("2010-01-08", "2012-06-08", "2015-04-09"),
    end_date = c("2010-02-08", "2012-07-08", "2015-05-09"))

ecat_scaled <- calculate_scaled_ecat(my_data)
}
\references{
Ryan, P.H., G.K. LeMasters, P. Biswas, L. Levin, S. Hu, M. Lindsey, D.I. Bernstein, J. Lockey, M. Villareal,
G.K. Khurana Hershey, and S.A. Grinshpun. 2007. "A Comparison of Proximity and Land Use Regression Traffic Exposure Models
and Wheezing in Infants." Environmental Health Perspectives 115(2): 278-284.  \url{https://doi.org/10.1289/ehp.9480}
}
